+++ utilities/helper_functions.py
@@ -1,3 +1,4 @@
+import copy
 import dataclasses
 import json
 import os
@@ -592,7 +593,7 @@ def replace_rich_text_user_channel(
     region_record: SlackSettings,
     client: WebClient,
     logger: Logger,
-) -> str:
+) -> Dict[str, Any]:
     """Replace user and channel ids with their names in a rich text block
 
     Args:
@@ -604,10 +605,12 @@ def replace_rich_text_user_channel(
     Returns:
         block (Dict[str, Any]): rich text block with slack ids replaced
     """
-    if not block or block.get("type") != "rich_text":
-        return block
+    new_block = copy.deepcopy(block)
 
-    for element in safe_get(block, "elements") or []:
+    if not new_block or new_block.get("type") != "rich_text":
+        return new_block
+
+    for element in safe_get(new_block, "elements") or []:
         if element["type"] in ["rich_text_section", "rich_text_preformatted", "rich_text_quote"]:
             for text in element["elements"]:
                 if text["type"] == "user":
@@ -620,6 +623,28 @@ def replace_rich_text_user_channel(
                     text["text"] = f"#{channel_name}"
                     text["type"] = "text"
                     del text["channel_id"]
+
+    return new_block
+
+
+def fix_from_llm_tags(block: Dict[str, Any]) -> Dict[str, Any]:
+    """Removes the 'from_llm' tag that Slack adds in the Android app to rich text blocks
+
+    Args:
+        block (Dict[str, Any]): rich text block with potential 'from_llm' tags
+
+    Returns:
+        Dict[str, Any]: rich text block with 'from_llm' tags removed
+    """
+
+    if not block or block.get("type") != "rich_text":
+        return block
+
+    for element in safe_get(block, "elements") or []:
+        if element["type"] in ["rich_text_section", "rich_text_preformatted", "rich_text_quote"]:
+            for text in element["elements"]:
+                if text.get("from_llm") is not None:
+                    del text["from_llm"]
 
     return block